lots of functionality: Spring Security, adding new plots, registering users with new user_roles

+++ tomatofarm/src/main/java/teksystems/tomatofarm/database/entity/Plot.java
@@ -17,11 +17,11 @@ public class Plot {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     //remember, plots are not to be created without an assigned user
     @Column(name = "user_id")
-    private Long userId;
+    private Integer userId;
 
     @Column(name = "soil_makeup")
     private String soilMakeup;
@@ -34,5 +34,4 @@ public class Plot {
 
     @Column(name = "spaces_taken")
     private Integer spacesTaken;
-
 }
